Handle undefined alerts from server (#13919)
In the event of a 422, it looks like the server returns an empty body, and causes the front-end to throw an error. This basically just leaves the VirtualHearingModal displaying when it should automatically close (I thought it might cause a white screen, but it doesn't). With this change, the modal closes on receiving an error.

I *think* the 422 is caused by an expired CSRF error.

https://sentry.ds.va.gov/department-of-veterans-affairs/caseflow/issues/9827/?referrer=webhooks_plugin

---

**Before:**

![Screen Recording 2020-04-07 at 3 35 50 PM](https://user-images.githubusercontent.com/1298274/78712312-70e3a600-78e6-11ea-92f5-daa0d1cb2171.gif)

---

**After:**

![Screen Recording 2020-04-07 at 3 36 20 PM](https://user-images.githubusercontent.com/1298274/78712240-4db8f680-78e6-11ea-8ad4-e26b007b2b67.gif)


### Description

- Added some handling around undefined alerts/body
- Lint changes