Fix pexip alias (#14099)
Resolves #14053

This PR adds a new column `alias_with_host` to future-proof how we generate virtual hearing links that are sent out to users and are displayed on `DailyDocket` and `Details` page. As far as we know, the pexip team will at some point have us use a different `client_host.`At that point we don't want any existing scheduled virtual hearings with the old `client_host` to break and not work. Therefore, if we always store and retrieve the alias from the database as opposed to formatting on the fly.  

### Description
- Create a new column `alias_with_host` in virtual hearings to store the alias in `BVA000000@client_host` format
- Use new column `alias_with_host` when generating links for emails and in `VirtualHearingLink` component
- fall back to formatting `alias` with `formatted_alias_with_host` if `alias_with_host` is nil
 
### Database Changes
*Only for Schema Changes*

 * [ ] Timestamps (created_at, updated_at) for new tables~
* [x] Column comments updated
* [x] Query profiling performed (eyeball Rails log, check bullet and fasterer output)
* [x] DB schema docs updated with `make docs`
 * [ ] Appropriate indexes added (especially for foreign keys, polymorphic columns, and unique constraints)~
* [x] #appeals-schema notified with summary and link to this PR