Do not filter disabilities (#14114)
connects #14074

### Description
When the feature toggle `disability_issue_test` is enabled, this removes the filters being applied to rating decisions (from disabilities) on a given decision review. It removes filters based on duplicates with decision issues as well as receipt dates.

This is being done for Martina to explore data in prod for specific cases, and is not intended to be in use in the application by users.