altered config to better support HMR: (#14245)
Connects #12738

### Description
We now wrap various React entry points with the `hot` HoC from `react-hot-loader`; this avoids the errors we would sometimes get during hot reload that would prevent rendering and redux rehydration. 

This covers these entry points:
- intake
- intakeEdit
- queue
- hearings

### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] Hot reload no longer errors out

### Testing Plan
This can be tested on a number of different pages, but here is an example:

1. As BVAEERDMAN (Ezra M Erdman), go to Your Queue and select one of the sample "vacate" decisions to draft
2. Select "Decision ready for review" from task actions
3. When looking at the "Review vacated decision issues" page, make a change to the title of the first item in `mtvCheckoutSteps.js`. 

In master branch, this causes an error that breaks the page. With this PR, things should smoothly update without losing place, etc.