Intake | Bug | Use saved withdrawn issue IDs when processing asynchronously (#14116)
connects #13483

### Description
This adds the ability for request issue withdrawals to fetch withdrawn issues from their IDs saved in the database on the request issues update. This is needed if they are being accessed asynchronously outside of the original update that was submitted which may happen if the VBMS `remove_contention` call fails on the first attempt.

There is a related [manual data fix GitHub issue](https://github.com/department-of-veterans-affairs/caseflow/issues/13399) to look for more instances of this and re-attempt processing for them.