removed `build:demo` npm script as it was unused; (#14316)
Connects #14022

### Description
This further refines the changes made in #14276, but moves the command to build storybook into the `build:production` script (`build:demo` apparently is never actually executed); relatedly, this also removes the `build:demo` script. This also moves the `@storybook/*` dependencies to normal as dev deps are cleansed in production environments.

### Acceptance Criteria
- [ ] Code compiles correctly