Revert "Intake | Bug Fix | Remove invalid characters in descriptions (#14247)" (#14359)
Reverts department-of-veterans-affairs/caseflow#14247

This fix was implemented due to an external bug impacting creating contentions with non-ASCII characters in the description.

An external fix for the bug was tested in UAT last week, and the plan is to deploy to prod this weekend, after testing on Sunday morning.  Not allowing non-ASCII characters causes some data corruption (contention text is used elsewhere as well), so once the external fix is in, we are going to revert our code as well to re-allow contentions with non-ASCII characters in the text.

Getting this PR ready so that if all goes well on Sunday we can shoot to have this merged in the Monday morning deployment.