Add schema for IntakesController#review (#14368)
Connects #14017 

### Description
This PR implements the version of this controller schema [tech spec](https://github.com/department-of-veterans-affairs/caseflow/pull/13326) which was discussed and approved by the Backend Working Group.

One significant improvement is included: the original proposal contained a `DocCompiler` class which walked the `Dry::Schema` AST and assembles a documentation-friendly object from the schema validator.

This approach turned out to be brittle in the face of dry-schema upgrades (from 1.4 to 1.5 in the past few months), so I swapped that with an abstraction layer (`ControllerSchema`) that's capable of easily generating both the `Dry::Schema` and documentation. Plus, it's much easier to read and understand than the AST parser, with just a couple tricky lines for programmatically using the Dry DSL.

The inaugural schema is on `IntakesController#review`, lives in the corresponding `app/schemas/intakes_schemas.rb`. The auto-documentation lives at `/route_docs` (pardon the hasty non-semantic HTML) and this schema can be found by searching for `intakes#review` on that page.

This schema has already paid a small dividend: suggesting a tweak in the frontend to treat `legacy_opt_in_approved` as a boolean wherever possible, rather than a string.

### Code Documentation Updates
- [ ] Add Caseflow wiki page on controller schema conventions