Add TimezoneService for address to timezone lookups (#13946)
Resolves #13889

### Description

  * Added new dependencies
    * [`countries`](https://github.com/hexorx/countries) - added to convert a country name into the [ISO 3166 2-letter code](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2) (which `tzinfo` can find a timezone for)
    * [`ziptz`](https://github.com/infused/ziptz) - maps US zip codes to timezones
    * `tzinfo` - explicitly declared this (it was being pulled in by another dependency before)
  * Added `TimezoneService` class, which can find the timezone for an `Address` instance
  * Added a _bunch_ of tests

Note: `tzinfo` and `ziptz` aren't 100% accurate. For example, a zip code in the U.S. Virgin Islands returns the timezone `"America/Puerto_Rico"`, but the [IANA timezone database](https://www.iana.org/time-zones) has one timezone for the U.S. Virgin Islands, [`"America/St_Thomas"`](https://time.is/U.S._Virgin_Islands).

I did some research into whether or not this would ultimately be an issue when displaying the hearing time, but it _seems_ like it should be fine because both timezones fall under the same geographic zone (which gets abbreviated as `AST` in the example below). There are other issues with the IANA timezone database, but since it appears to be the standard for classifying timezones, it's probably best to use.

```
[18] pry(main)> TZInfo::Timezone.get("America/Puerto_Rico").strftime("%A, %-d %B %Y at %-l:%M%P %Z", Time.now.utc)
=> "Monday, 13 April 2020 at 1:57pm AST"
[19] pry(main)> TZInfo::Timezone.get("America/St_Thomas").strftime("%A, %-d %B %Y at %-l:%M%P %Z", Time.now.utc)
=> "Monday, 13 April 2020 at 1:58pm AST"
```