Resuce RecordNotUniqueError in UpdateCachedAppealsAttributesJob (#14606)
### Description
`UpdateCachedAppealsAttributesJob` has been failing lately due to `ActiveRecord::RecordNotUniqueError` being thrown when trying to create `BgsPowerOfAttorney` object. From the [sentry event](https://sentry.ds.va.gov/department-of-veterans-affairs/caseflow/issues/10307/?referrer=webhooks_plugin), it looks like this is because of a race-condition where another process is trying to create the same object. So in order to let the job continue to the next appeal without breaking, we can resuce and return the object if it exists. 

[Slack thread for reference](https://dsva.slack.com/archives/CN3FQR4A1/p1593032872085600)