Make controller schemas stricter by discarding unknown fields (#14628)
### Description
In the BEWG discussion on controller schemas, we talked about a few approaches to strictness that would encourage keeping the docs up-to-date with the code -- i.e. if you want to start using a new field, you must update the schema first. For the sake of small PRs, strictness wasn't included in the initial implementation.

The version of strictness here just discards any unknown fields before passing params to the rest of the controller. It's a more lenient alternative to failing hard, so that browser users or middleware can tack on something like `&cache_buster=1234` to the query string, without being rejected by the validation.

### Testing Plan
1. Brand new spec for ControllerSchema!