Reinstate db:migration:redo check in CI (#14696)
### Description
The db:migration:redo rake task is shorthand for rolling back the latest migration, and then re-applying it. This allows us to check for syntax errors and other loudly-failing migration-time errors in CI. It was removed in #13768 which squashed away all migrations at the time, and we forgot to reinstate it.