Add request throttling to VA.gov service (#14695)
Rate limiting was removed in https://github.com/department-of-veterans-affairs/caseflow/pull/12366, and caching was put in instead. Unfortunately, we are still running into a rate limit everytime we run the `FetchHearingLocationsForVeteransJob` job:

<img width="1373" alt="Screen Shot 2020-07-14 at 3 09 27 PM" src="https://user-images.githubusercontent.com/1298274/87466465-4a6b9000-c5e4-11ea-918e-3fd7ce2e76e2.png">

One complaint in #12366 was that the sleep call in `FetchHearingLocationsForVeteransJob` was related to a degradation in performance in the seed file. Moving the sleep call into the non-fake `ExternalApi::VADotGovService`, means that the sleep call won't be triggered in the dev environment.

### Description

 - Add request throttling to VA.gov service