Intake | G-code | Add code for opting in G code appeals (#14731)
connects #14574

### Description
This adds code to appeal-repository to close appeals with the "G" code disposition. We are actually running the same code for closing undecided appeals (confirmed with Jed Wagner that this is okay and preferable to cherry picking the fields).  Due to the naming and because we may get more similar requests in the future, I broke this out into a method named for opt-ing in already decided appeals.

### Acceptance Criteria
- [x] Add a method for opting in "decided" legacy appeals to AppealRepository
- [x] The code can be the same as AppealRepository.close_undecided_appeal!, but allows legacy appeals decided with the "G" disposition to be "re-closed"
- [x] Update LegacyAppeal.close_single to use the new method added when the current disposition is "G"

### Testing Plan
- [x] Add model tests for opting in all issues on a legacy appeal that is already closed with the "G" disposition and checking that it gets a new disposition of "O"

### User Facing Changes
None

### Code Documentation Updates
- [x] Add or update code comments at the top of the class, module, and/or component.

### Storybook Story
N/A

### Database Changes
N/A