Restructured Jest Commands (#14836)
### Description
This restructures and extends the yarn scripts related to Jest. It removes the `-u` flag from the default `test:jest` command, as that seemed to prevent CircleCI from catching snapshot mismatches. It also adds `test:jest:watch` and `test:jest:update` commands (the latter allows the user to update snapshots as the prior default command did).

### Acceptance Criteria
- [ ] Code compiles correctly