Add structured renderer for intake data models (#14843)
### Description
This PR adds a utility class and a monkey-patchable module to `lib/helpers/` for extending @tomas-nava's task tree structure renderer to intake data models.

The difference and biggest challenge here is that intake data models are not uniformly recursive like task trees. However, there is somewhat of a common hierarchy to intake, and by using a framework that delegates structure-discovery to each model, we can build a renderable tree nevertheless.

This was motivated by a BEWG [thread](https://dsva.slack.com/archives/CQGMLMWFK/p1590076450029800) on where to put utility methods -- not invoked by the main codebase, but available in prod console to 🦇 team. My intention is that `lib/helpers/` can contain slightly more experimental code like this, so that we can still properly review it for safety.

### Usage
In the Rails console,
```
> vet = Veteran.find(veteran_id)  # see RENDERABLE_CLASSNAMES for supported types
> puts IntakeRenderer.render(vet)
Veteran 1 (PID: 500000000)
├── Appeal 1 (received 2020-07-01, 8a015db0-72ad-4438-ab09-d43f7e50dc4e)
│   ├── VeteranClaimant 1 (PID: 500000000)
...

> puts IntakeRenderer.render(vet, show_pii: true)  # show_pii defaults to false, of course
Veteran 1 (Bob Smith, FN: 500000000, PID: 500000000)
├── Appeal 1 (received 2020-07-01, 8a015db0-72ad-4438-ab09-d43f7e50dc4e)
│   ├── VeteranClaimant 1 (Bob Smith, PID: 500000000)
...

> IntakeRenderer.patch_intake_classes  # you can also monkey-patch the supported classes
> vet.render_intake(show_pii: true)
... same as above
```