Removed default of `true` for `isClearable` in `SearchableDropdown` (#15300)
### Description
A previous PR changed the default for the `isClearable` prop on `SearchableDropdown` to `true`. This was done with the approval of design, but it had unintended consequences. There are a number of components that utilize `SearchableDropdown` that don't anticipate a null/empty state. This often results in JS errors due to callback functions that attempt to access objects but instead receive `null` arguments. 

This PR removes the default prop from `SearchableDropdown`, but adds comments to the proptypes to indicate that `isClearable` should be preferred unless it would not be relevant in the specific use case.

### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] Default for `isClearable` is false/undefined
- [ ] Jest snapshots are updated to reflect the change