Do not add CO as a location to geomatch from if veteran lives in VA or MD (#15418)
Resolves #15415

### Description
We discovered that there was a stale logic adding CO to locations to geomatch from if Veteran lives in VA or MD. If CO ends up becoming the closest location to Veteran then they are geomatched to central office but the central queue has logic to only display appeals which have specifically requested central office hearing. In other words, if veterans who did not request central office hearings but are geomatched to central office, they would not show up in central queue. This PR removes that stale logic per permission from the hearing branch.