[Part 1] Ensure that a new VACOLS hearing is created when postponing and rescheduling a hearing (#15460)
Resolves #15448 

Related: https://github.com/department-of-veterans-affairs/caseflow/pull/11144
Also this PR notes the same behavior we are fixing: https://github.com/department-of-veterans-affairs/caseflow/pull/5974

### Description

This resolves a very old bug that was causing the `hearing_pkseq` ID to get incremented twice on each insert.

```
[404739,
 404741,
 404743,
 404745,
 404747]
```

The double-increment issue was due to a trigger in the VACOLS database that *also* increments the primary key on inserts:

https://github.com/department-of-veterans-affairs/caseflow/blob/c5964822ba99d051a321d3d50ba8db54356f88ea/local/vacols/vacols_copy_4_triggers_dev.sql#L27-L37

Rails will by default increment the primary key, and, as a result, IDs in the sequence were being skipped when inserting new data:

```
[1] pry(main)> VACOLS::CaseHearing.create!(hearing_type: 'V', hearing_date: Date.today)

# Rails grabbing the next sequence value
[2020-10-20 14:24:49 -0400]   next sequence value (5.2ms)  SELECT "HEARSCHED_PKSEQ".NEXTVAL FROM dual

# Rails populating `hearing_pkseq` on insert
[2020-10-20 14:24:49 -0400]   VACOLS::CaseHearing Create (4.4ms)  INSERT INTO "HEARSCHED" ("HEARING_TYPE", "HEARING_DATE", "HEARING_PKSEQ") VALUES (:a1, :a2, :a3)  [["hearing_type", "V"], ["hearing_date", "2020-10-20 00:00:00"], ["hearing_pkseq", 404825]]

# Returned value shows the ID that Rails populated
=> #<VACOLS::CaseHearing:0x0000565050411120
 hearing_pkseq: 404825,
...>

# Trying to find by that ID causes a RecordNotFound error
[2] pry(main)> VACOLS::CaseHearing.find(404825)
[2020-10-20 14:27:06 -0400]   VACOLS::CaseHearing Load (1.4ms)  SELECT  "HEARSCHED".* FROM "HEARSCHED" WHERE "HEARSCHED"."HEARING_PKSEQ" = :a1 FETCH FIRST :a2 ROWS ONLY  [["hearing_pkseq", 404825], ["LIMIT", 1]]
ActiveRecord::RecordNotFound: Couldn't find VACOLS::CaseHearing with 'hearing_pkseq'=404825

# Getting the last created hearing shows a `hearing_pkseq` value that is 1 greater than what was inserted
[3] pry(main)> VACOLS::CaseHearing.last
[2020-10-20 14:27:31 -0400]   VACOLS::CaseHearing Load (2.2ms)  SELECT  "HEARSCHED".* FROM "HEARSCHED" ORDER BY "HEARSCHED"."HEARING_PKSEQ" DESC FETCH FIRST :a1 ROWS ONLY  [["LIMIT", 1]]
=> #<VACOLS::CaseHearing:0x0000565055e36b00
 hearing_pkseq: 404826,
...>
```

There were a number of ways that developers worked around this, including this call here, which was intended to find the most recently created `VACOLS::CaseHearing` after doing an insert:

```
vacols_record = VACOLS::CaseHearing.for_appeal(appeal.vacols_id).where(vdkey: hearing_day.id)
         .order(addtime: :desc).last
```

This was a fix for #10643, but the issue still persisted until now. We are seeing different `HearingTasks` that point to the same legacy hearing.

Once we fixed the double-increment problem, we were able to successfully load the case hearing by ID after we create it, and properly associate it with the new hearing task. This suggests that there might be `VACOLS::CaseHearing` *that are not linked to a hearing task, and can be deleted*.

Note: Fixing this issue caused 2 subsequent issues to show up, which we are resolving via. follow up PRs.

### Acceptance Criteria
- [x] Ensure that duplicates aren't created when working the assign hearing disposition task for virtual hearings
- [x] Find out the scope of this issue by looking for hearing tasks with the same hearing


### Code Documentation Updates
- [x] Add or update code comments at the top of the class, module, and/or component.