Handle updating correction type in EndProductUpdate (#15490)
Connects #15128 

### Description
After an edit is submitted, if the the correction type changes, this code will update the correction type on the request issues.

## Acceptance criteria
- [x] Add a method to EndProductUpdate to perform the data changes related to the EP code change
- [x] If the code change includes a correction type change
  - [x] Calculate the new correction type from the new code (use EP code map)
  - [x] Update all of the active request issues on the end product establishment in Caseflow DB
    - [x] Update correction_type on the request issues to the new correction type
- [x] Add EndProductUpdate factory
- [x] Add spec

### Testing Plan
1. Added a new spec for EPU.

This PR is just a model update, so there doesn't need to be any additional tests. But I invite reviewers to scrutinize the new model test, since it lays out the structure for all the other types of EPUs in the future.