Api v3 hlr json schema (#15492)
Resolves #15491 

### Description
Currently the create endpoint for API V3 HLR uses a hand rolled validation to check all the types and paths in the passed in JSON body. The PR for this issue will remove the current validation and replace it with JSON Schemer validation. This provides a cleaner approach to validating and removes a lot of custom validation methods and files.

### Acceptance Criteria
 - [x] Replace current validation with JSON Schemer
 - [x] Update all broken specs
 - [x] Remove obsolete methods and files