Write comments for Build Hearing Schedule functionality (#15623)
Resolves https://vajira.max.gov/browse/CASEFLOW-227

This PR is a result of research on Build Hearing Schedule functionality. The comments are my understanding of the algorithm and is meant to help guide developers to understand the algo better. The write up for my research lives [here](https://hackmd.io/JJfSpuSHTAWxJo2ebC9STw) and [here](https://github.com/department-of-veterans-affairs/caseflow/wiki/Caseflow-Hearings#build-hearing-schedule)