Add fields to IDT appeals detail serializer (#15633)
### Description
Data for attorneys in the CorpDB isn't always formatted the same as other claimants. The first and last names are missing, but full name is available. Sometimes the full name is what's in address_line_1, and the rest of the address is in address_line_3, which wasn't being sent to IDT before.

[Relevant Slack thread from Paul Saindon](https://dsva.slack.com/archives/CD5DAQNCU/p1605898805003600)

### Acceptance Criteria
- [x] Code compiles correctly

### Testing Plan
1. Paul tested in UAT
1. Added small model test just for new functionality

- [x] For higher-risk changes: [Deploy the custom branch to UAT to test](https://github.com/department-of-veterans-affairs/appeals-deployment/wiki/Applications---Deploy-Custom-Branch-to-UAT)