Rerun stuck jobs before reporting; show additional details on Slack (#15776)
[Related Slack Thread](https://dsva.slack.com/archives/C3EAF3Q15/p1610471973156500?thread_ts=1610447169.151600&cid=C3EAF3Q15)

This PR reruns the `CreateConferenceJob` in `StuckVirtualHearingsChecker` for all virtual hearings where `conference_id` is `nil` before reporting the stuck virtual hearings. It also tweaks the report to show the css_id of user who updated the virtual hearing, uuid/vacols_id of hearing, and a link to a document where we're tracking the virtual hearings that are stuck.