class << self the whole class (#15797)
### Description
Every single method in here was a class method. It bugged me that they all started with `self.`, so I `class << self`ed it.

This is not a sprint story; just a minor pet peeve I discovered while working on something else.

### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] You think this is a worthwhile change

### Testing Plan
There should be no feature change, and I trust our test suite to confirm that. You are free to disagree.