Create dependent records and appeals for CAVC appeals FactoryBot (#15811)
Bumps [CASEFLOW-93](https://vajira.max.gov/browse/CASEFLOW-93) and Part 2 of #15774.

### Description
Make `cavc_remand` FactoryBot for robust.

An appeal resulting from a CAVC Remand depends on the `CavcRemand` record to exist so that the appropriate tasks can be created depending on the `CavcRemand.subtype`. It also relies on records about the claimant, veteran, decision issues, and decision documents of the source appeal of the `CavcRemand`.

When creating a CAVC Remand appeal using FactoryBot, the factory should allow for the dependencies to be passed in or created as needed. In the end, all the records should be consistent with each other, at least enough to be used in an RSpec test. 

### Acceptance Criteria
- [x] Code compiles correctly