Adding fake rooms to build Q4 FY21 hearing days (#15993)
### Description
Adding fake rooms to Caseflow again to get back up to 28. The Branch needs this many rooms to build 1,705 hearing days for Q4 FY21, and the new algorithm for building hearing days without rooms has [an emerging requirement](https://vajira.max.gov/browse/CASEFLOW-1078) we're not able to finish in time for them to stay on schedule.