[Fix flaky] Editing Virtual Hearings from Hearing Details - cannot detect expected alert (#15988)
Resolves #15983

### Description
The test in `spec/feature/hearings/virtual_hearings/hearing_details_spec.rb` is failing because it expects a progress alert but the output is a success alert. This is because the progress alert gets replaced by a success alert when some data changes (after polling). It appears that in CircleCi unlike dev success alert is too quick to replace progress alert.
 I also made other changes in the file so that we're always expecting the success alert. 

### Testing Plan
This one is hard to reproduce as it doesn't seem to fail at all locally.