Fix median, and add test coverage (#16008)
Also fixes the old test that asserted incorrect behavior!

Fixes #16006

### Description
We ran into [this weird flake](https://app.circleci.com/pipelines/github/department-of-veterans-affairs/caseflow/39962/workflows/7f06f229-5ed6-4922-b58b-82d2241520b5/jobs/152724) recently.

I _think_ the immediate cause of the error was that we assumed the database output was in chronological order, but never used an `ORDER BY`, so there was no contract for the ordering.

But in looking at this, I was skeptical that `Reporter#median` was correct. I added some tests that prove it is--as long as the input sorted. Since that was clearly not always true, I made the method sort its input, since it returns nonsense otherwise. (Previously, the output was literally the middle value, so the "median" of `[1, 1, -10000, 1, 1]` would be `-10000` which is certainly not what statisticians would expect.)

"Fixing" the method broke the existing test (!), so I also fixed that.

### Acceptance Criteria
- [ ] Tests pass successfully