Remove tied to VLJ Requirement for Non-Legacy Hearing Cases (#16016)
Resolves [CASEFLOW-1225](https://vajira.max.gov/browse/CASEFLOW-1225)

As a VLJ, I need the "tied to VLJ" requirement to be removed for any case that is not a legacy hearing case, as there is no legal requirement for any case other than a legacy hearing case to be assigned to me. 

### Description
Changed the `SELECT_PRIORITY_APPEALS` query to not populate the VLJ column with a judge id for priority legacy cases that have "PRIOR_DECISIONS" by a judge. The VLJ column determines if the case is tied to a judge. The effect is that these priority legacy cases decided by a judge (aka, post-remand cases) will no longer be distributed when requesting cases tied to a judge.

The VLJ column will still be populated for priority legacy cases that have a hearing.

### Acceptance Criteria
- [ ] Code compiles correctly

### Testing Plan
After Monday morning's priority push job (and after this PR is merged), check the distribution results in #appeals-job-alerts or at https://query.prod.appeals.va.gov/dashboard/7. No priority legacy post-remand cases should be distributed when requesting cases tied to a judge.