Jest Tests & Story for ColocatedTaskListView (#16009)
Connects 15764

### Description
This updates tests (moving from Karma to Jest) for `ColocatedTaskListView` component, and also adds Storybook stories for same. 

This should fix the flakey test that was caused by the use of Moment-based dates that were subject to variable behavior depending on time of day when the test was run.


### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] Tests for `ColocatedTaskListView` are unaffected by system time when run
- [ ] Storybook story exists