Adds basic test for completed_hearing_on_previous_appeal? (#16028)
### Description

We questioned whether this method might _always_ return true.
I wrote a quick test that found that is not the case.

I welcome feedback on whether this test is worth committing. It was a quick attempt at TDD that ended up showing the bug I was investigating was not present. Adding additional test cases here is not a priority right now.

### Acceptance Criteria
- [x] Code compiles correctly
- [x] You believe the test adds value