Add all possible US states/territories (#16079)
I noticed that outside of the 50 states, only DC and PR were included in the states constant. This PR (heh) adds the remaining options from this very official [USPS page](https://faq.usps.com/s/article/What-are-the-USPS-abbreviations-for-U-S-states-and-territories).

`STATES` is only used by the `AddressForm` component, and that component is only used on the `AddClaimantForm` and `AddPoaPage` components (which matches my understanding that no other part of Caseflow intakes addresses).