Sets role="button" on CopyTextButton
Context: I do not actually think this makes a lot of sense. We
have an accessibility ticket stating that this element does not
have a discoverable role. While the 'role' attribute was not
present, setting role="button" on a literal button does not seem
like it would help anyone. JAWS and VoiceOver identify the button
as such with or without this change.

Pushing this PR up because it implements what is requested, but I
am hoping to have a conversation about whether the ticket may be
obsolete or perhaps incorrectly describing a problem.