Hide early morning slots and slots before scheduled hearings (#16130)
### Description
This changes only the time slots that are hidden. This resolves two of the issues with the timeslot prototype:
1. Show only slots that are after 8:30am local time.
2. Edit: This change was already merged by adding 'abs' to `const slotFull = hearingTimes.some...`. Don't show slots that are less than an hour before a scheduled hearing. Ex: 10:45am hearing hides a 10:30am slot.


### Background
This is a small, much safer subset of this PR: https://github.com/department-of-veterans-affairs/caseflow/pull/16070

The work in that linked PR that is not included here resolves a variety of issues with the custom time dropdown. I believe that work was very useful in allowing user sessions, but I'm not sure it needs to be merged.
- We are planning to replace that custom time dropdown.
- Those changes would potentially also impact the daily docket, assign hearing, and hearing conversion.