Display More Information About HearingDay When Scheduling Veteran (#16086)
Resolves CASEFLOW-395

### Description
Provides more information about hearing days on the "Schedule Veteran" screen. in the left hand column.

### Test Plan
- Go to the [Schedule Veterans -> Assign Hearings page](http://localhost:3000/hearings/schedule/assign?regional_office_key=RO17&tab=unassignedTab) for St. Petersburg, FL.
- Verify that the list under the "Available Hearing Days" is in the new format (see After screenshot below).
- Verify that clicking on the different hearing days changes which one is selected and updates the list of Scheduled Veterans on the right hand side of the page.
- Verify that the information in the left hand side list is correct. It should be:
```
[Date of Hearing Day]                [X of Y]
[HearingDay Type] - [Judge/Room]  [scheduled]
```

Note: The screen jumps around when the scrollbar for the entire page appears/disappears. This is existing behavior all over caseflow and is not possible to fix as part of this work.

Also note: The 'available slots' information displayed comes from HearingDay.total_slots, which is also displayed as "(Y slots remaining)" on this page. This is not the number of TimeSlot(s) that would be displayed if you are looking at the TimeSlot prototype work.

#### Before
![Screen Shot 2021-04-02 at 1 36 40 PM](https://user-images.githubusercontent.com/6818839/113439522-bda67380-93b8-11eb-9fd3-dbe4eb85e923.png)

#### After
![Screen Shot 2021-04-02 at 1 35 52 PM](https://user-images.githubusercontent.com/6818839/113439546-ca2acc00-93b8-11eb-8fad-66dbb3aa0614.png)


### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] Information shown is correct.
- [ ] Code moved to 'util.js' doesn't change output of other components.